feat: add preparation scene
